refactor: migrate app action consumers
