+++ src/workspaces/intune/NewIntuneWorkspace.tsx
@@ -15,7 +15,7 @@ import {
 import { getLogListMetrics } from "../../lib/log-accessibility";
 import { useIntuneStore } from "./intune-store";
 import { useUiStore } from "../../stores/ui-store";
-import { useAppActions } from "../../components/layout/Toolbar";
+import { useAppActions } from "../../hooks/use-app-actions";
 import { EventLogSurface } from "./EventLogSurface";
 import { InvestigationPanel } from "./InvestigationPanel";
 import { OverviewSurface } from "./OverviewSurface";
